+++ docs/CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog
 - Enh #30: Tabulator Support
 - Fix #50: Right arrow navigation not working in some cases
 - Enh #102: Source Button Position
+- Enh #104: Source Mode, Text Area Style
 
 
 1.1.3 (Unreleased)